collisionballs better + documentation out of the way (can be recovered!)

+++ shaders/2d_flat.fs
@@ -1,4 +1,4 @@
-precision highp float;
+precision mediump float;
 
 uniform sampler2D tex0;
 uniform vec3 unib[3];
@@ -16,8 +16,8 @@ void main(void) {
   coord.y = unif[15][2] - coord.y; // top left convension though means flipping image!
   if (coord.x <= unif[14][0] || coord.x > unif[14][0]+unif[15][0] ||
       coord.y <= unif[14][1] || coord.y > unif[14][1]+unif[15][1]) discard; // only draw the image once
-  coord -= unif[14].xy;
-  coord *=  pix_inv;
+  coord -= unif[14].xy; // offset
+  coord *=  pix_inv; // really dividing to scale 0-1 i.e. (x/w, y/h)
   vec4 texc = texture2D(tex0, coord);
   if (texc.a < unib[0][2]) discard; // ------ to allow rendering behind the transparent parts of this object
   gl_FragColor = texc;
